LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsViewsScreenNavigationCategory to ObjectDefinitionsViewsScreenNavigationEntry, implement getCategoryKey in ObjectDefinitionsViewsScreenNavigationEntry and make ObjectDefinitionsViewsScreenNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
